Create Tic Tac Toe Part 3
In this part I have added a 2 player feature along with the option to play with the AI. I have created two AI modes: easy or hard. Hard mode will never lose(probably). I tried my best to make this as interactive as possible. Will be considering adding new features in the future.